requirements.txt
dora_search
colorlog
numpy
torch
